+++ src/index.ts
@@ -33,6 +33,7 @@ export * from "./io/completion.js";
 export * from "./io/editor.js";
 export * from "./io/history.js";
 export * from "./io/keybindings.js";
+export * from "./io/render.js";
 export * from "./io/safe-path.js";
 export * from "./input.js";
 export * from "./models/catalog.js";